On 64-bit architectures, use long long for 64-bit ints

+++ cko_types.h
@@ -4,12 +4,14 @@
 typedef unsigned char cko_u8;
 typedef unsigned short cko_u16;
 typedef unsigned int cko_u32;
-typedef unsigned long cko_u64;
+//typedef unsigned long cko_u64;
+typedef unsigned long long cko_u64;
 
 typedef char cko_s8;
 typedef short cko_s16;
 typedef int cko_s32;
-typedef long cko_s64;
+//typedef long cko_s64;
+typedef long long cko_s64;
 
 int cko_types_test();
